+++ components/uni-fav/uni-fav.vue
@@ -8,7 +8,7 @@
 		<!-- #ifndef MP-ALIPAY -->
 		<uni-icons :color="fgColor" :style="{color: checked ? fgColorChecked : fgColor}" class="uni-fav-star" size="14" type="star-filled" v-if="!checked && (star === true || star === 'true')" />
 		<!-- #endif -->
-		<text :style="{color: checked ? fgColorChecked : fgColor}" class="uni-fav-text">{{ checked ? contentText.contentFav : contentText.contentDefault }}</text>
+		<text :style="{color: checked ? fgColorChecked : fgColor}" class="uni-fav-text">{{ checked ? contentFav : contentDefault }}</text>
 	</view>
 </template>
 
@@ -28,6 +28,15 @@
 	 * @event {Function} click 点击 fav按钮触发事件
 	 * @example <uni-fav :checked="true"/>
 	 */
+
+	import {
+		initVueI18n
+	} from '@dcloudio/uni-i18n'
+	import messages from './i18n/index.js'
+	const {
+		t
+	} = initVueI18n(messages)
+
 	export default {
 		name: "UniFav",
 		// TODO 兼容 vue3，需要注册事件
@@ -65,11 +74,19 @@
 				type: Object,
 				default () {
 					return {
-						contentDefault: "收藏",
-						contentFav: "已收藏"
+						contentDefault: "",
+						contentFav: ""
 					};
 				}
 			}
+		},
+		computed: {
+			contentDefault() {
+				return this.contentText.contentDefault || t("uni-fav.collect")
+			},
+			contentFav() {
+				return this.contentText.contentFav || t("uni-fav.collected")
+			},
 		},
 		watch: {
 			checked() {