+++ components/uni-list-item/uni-list-item.vue
@@ -214,7 +214,7 @@
 				}
 			},
 			pageApi(api) {
-				uni[api]({
+				let callback = {
 					url: this.to,
 					success: res => {
 						this.$emit('click', {
@@ -225,9 +225,24 @@
 						this.$emit('click', {
 							data: err
 						});
-						console.error(err.errMsg);
 					}
-				});
+				}
+				switch (api) {
+					case 'navigateTo':
+						uni.navigateTo(callback)
+						break
+					case 'redirectTo':
+						uni.redirectTo(callback)
+						break
+					case 'reLaunch':
+						uni.reLaunch(callback)
+						break
+					case 'switchTab':
+						uni.switchTab(callback)
+						break
+					default:
+						uni.navigateTo(callback)
+				}
 			}
 		}
 	};