+++ components/uni-swipe-action-item/mpalipay.js
@@ -25,14 +25,6 @@ export default {
 			this.swipeaction.children.push(this)
 		}
 	},
-
-	// beforeDestroy() {
-	// 	this.swipeaction.children.forEach((item, index) => {
-	// 		if (item === this) {
-	// 			this.swipeaction.children.splice(index, 1)
-	// 		}
-	// 	})
-	// },
 	mounted() {
 		this.isopen = false
 		setTimeout(() => {
@@ -62,13 +54,6 @@ export default {
 				})
 			}
 		},
-		// onClick(index, item, position) {
-		// 	this.$emit('click', {
-		// 		content: item,
-		// 		index,
-		// 		position
-		// 	})
-		// },
 		/**
 		 * 移动触发
 		 * @param {Object} e