+++ components/uni-swipe-action-item/mpwxs.js
@@ -1,32 +1,14 @@
+// #ifdef APP-VUE|| MP-WEIXIN || H5
 import { isPC } from "./isPC"
 export default {
 	data() {
 		return {
-			position: [],
-			button: {},
-			btn: "[]"
+			is_show:'none'
 		}
 	},
 	watch: {
-		button: {
-			handler(newVal) {
-				this.btn = JSON.stringify(newVal)
-			},
-			deep: true
-		},
-		show(newVal) {
-			if (this.autoClose) return
-			if (!this.button) {
-				this.init()
-				return
-			}
-			this.button.show = newVal
-		},
-		leftOptions() {
-			this.init()
-		},
-		rightOptions() {
-			this.init()
+		show(newVal){
+			this.is_show = this.show
 		}
 	},
 	created() {
@@ -35,20 +17,11 @@ export default {
 			this.swipeaction.children.push(this)
 		}
 	},
-	mounted() {
-		this.init()
-	},
-	// fixme by mehaotian 在页面激活的时候需要重新获取元素信息
-	activated(){
-		this.init()
+	mounted(){
+		this.is_show = this.show
 	},
 	methods: {
-		init() {
-			clearTimeout(this.swipetimer)
-			this.swipetimer = setTimeout(() => {
-				this.getButtonSize()
-			}, 50)
-		},
+		// wxs 中调用
 		closeSwipe(e) {
 			if (!this.autoClose) return
 			this.swipeaction.closeOther(this)
@@ -56,11 +29,9 @@ export default {
 
 		change(e) {
 			this.$emit('change', e.open)
-			let show = this.button.show
-			if (show !== e.open) {
-				this.button.show = e.open
+			if (this.is_show !== e.open) {
+				this.is_show = e.open
 			}
-
 		},
 
 		appTouchStart(e) {
@@ -100,24 +71,11 @@ export default {
 				position
 			})
 			// #endif
-		},
-		getButtonSize() {
-			const views = uni.createSelectorQuery().in(this)
-			views
-				.selectAll('.uni-swipe_button-group')
-				.boundingClientRect(data => {
-					let show = 'none'
-					if (this.autoClose) {
-						show = 'none'
-					} else {
-						show = this.show
-					}
-					this.button = {
-						data,
-						show
-					}
-				})
-				.exec()
 		}
 	}
 }
+
+// #endif
+// #ifndef APP-VUE|| MP-WEIXIN || H5
+export default {}
+// #endif