+++ components/uni-tag/uni-tag.vue
@@ -89,7 +89,7 @@
 					isTrue(circle) ? 'uni-tag--circle' : '',
 					isTrue(mark) ? 'uni-tag--mark' : '',
 					'uni-tag--' + size,
-					type === 'default' ? 'uni-tag--default' : 'uni-tag-text',
+					// type === 'default' ? 'uni-tag--default' : 'uni-tag-text',
 					isTrue(inverted) ? 'uni-tag-text--' + type : '',
 					size === 'small' ? 'uni-tag-text--small' : ''
 				]