+++ changelog.md
@@ -1,3 +1,36 @@
+## 3.2.4（2021-09-07）
+- 修复 vue3 在 H5 编译报错的 bug
+- 新增 同步 uni-ui
+	- 新增 uni-ui 组件支持国际化 i18n
+	- uni-data-checkbox 修复 在uni-forms中 modelValue 中不存在当前字段，当前字段必填写也不参与校验的问题
+	- uni-datetime-picker 优化 取消选中时（范围选）直接开始下一次选择, 避免多点一次
+	- uni-datetime-picker 优化 移动端支持清除按钮，同时支持通过 ref 调用组件的 clear 方法
+	- uni-datetime-picker 优化 调整字号大小，美化日历界面
+	- uni-datetime-picker 修复 因国际化导致的 placeholder 失效的 bug
+	- uni-file-picker 修复 return-type="object" 时且存在v-model时，无法删除文件的Bug
+	- uni-file-picker 新增 参数中返回 fileID 字段
+	- uni-file-picker 修复 腾讯云传入fileID 不能回显的bug
+	- uni-file-picker 修复 选择图片后，不能放大的问题
+	- uni-link 修复 在 nvue 下不显示的 bug
+	- uni-list 修复 在vue3中to属性在发行应用的时候报错的bug
+	- uni-search-bar 修复 value 属性与 modelValue 属性不兼容的Bug
+	- uni-swipe-action 优化 close-all 方法
+	- uni-collapse 优化 show-arrow 属性默认为true
+	- uni-collapse 新增 show-arrow 属性，控制是否显示右侧箭头
+	- uni-data-checkbox 修复 单选 list 模式下 ，icon 为 left 时，选中图标不显示的问题
+	- uni-easyinput 修复 在 uni-forms 的动态表单中默认值校验不通过的 bug
+	- uni-file-picker 修复 由于 0.2.11 版本引起的不能回显图片的Bug
+	- uni-file-picker 新增 clearFiles(index) 方法，可以手动删除指定文件
+	- uni-file-picker 修复 v-model 值设为 null 报错的Bug
+	- uni-swipe-action 新增 close-all 方法，关闭所有已打开的组件
+	- uni-swipe-action 新增 resize() 方法，在非微信小程序、h5、app-vue端出现不能滑动的问题的时候，重置组件
+	- uni-swipe-action 修复 app 端偶尔出现类似 Page[x][-x,xx;-x,xx,x,x-x] 的问题 
+	- uni-swipe-action 优化 微信小程序、h5、app-vue 滑动逻辑，避免出现动态新增组件后不能滑动的问题
+
+	
+## 3.2.3（2021-08-27）
+- 优化 tabbar 页面移除 vuex 相关代码
+- 新增 适配 vue3 (app)
 ## 3.2.2（2021-08-10）
 - 新增 适配快手小程序
 - 新增 同步 uni-ui